==> Code/05_test_real_data_cv.R <==
### Install packages ####
install_local <- F
library(devtools)

if (install_local) devtools::install_local("/Users/juliaindivero/Library/CloudStorage/Dropbox/sdmTMB-newlogistic.zip")
if (!install_local) remotes::install_github("pbs-assess/sdmTMB", dependencies = TRUE, ref="newlogistic", force=T)
library(sdmTMB)

### load helper functions ####
source("Code/util_funs.R")

### Set ggplot themes ###
theme_set(theme_bw(base_size = 25))
theme_update(panel.grid.major = element_blank(), panel.grid.minor = element_blank())

### Load Data ####
sci_name <- "Anoplopoma fimbria" #"Anoplopoma fimbria" # "Sebastolobus altivelis"   #"Eopsetta jordani" 
spc <- "sablefish" #"longspine thornyhead" # "petrale sole" #"sablefish"
dat.by.size <- length_expand(sci_name)
dat <- load_data(spc = spc, dat.by.size = dat.by.size)

#Constrain depth?
constrain_depth <- F
if(constrain_depth) dat <- subset(dat, depth<600)

#Constrain depth for petrale
constrain_depth <- F
if(constrain_depth) dat <- subset(dat, depth<500)

## Scale, rename, and calculate variables ##
dat$temp_s <- (scale(dat$temp))
dat$po2_s <- (scale(dat$po2))
dat$log_depth_scaled <- scale(log(dat$depth))
dat$log_depth_scaled2 <- with(dat, log_depth_scaled ^ 2)
dat$jday_scaled <- scale(dat$julian_day)
dat$jday_scaled2 <- with(dat, jday_scaled ^ 2)
dat$X <- dat$longitude
dat$Y <- dat$latitude
dat$cpue_kg_km2 <- dat$cpue_kg_km2 * (dat$p2+dat$p3)
dat$year <- as.factor(dat$year)

# Remove outliers = catch > 10 sd above the mean
dat$cpue_s <- scale(dat$cpue_kg_km2)
dat <- dplyr::filter(dat, cpue_s <=20)
### Calculate inverse temp ####
kelvin = 273.15 #To convert to Kelvin
boltz = 0.000086173324 #Boltzman's constant
tref <- 7 #Reference temperature in celsius
dat$invtemp <- (1 / boltz)  * ( 1 / (dat$temp + 273.15) - 1 / (tref + 273.15)) #invtemp 

### Make mesh ####
mesh <- make_mesh(dat, xy_cols = c("X", "Y"), n_knots = 250)

## Get initial values ##                                                                                                
init_vals <- get_inits()

### Plot data vs. depth and po2 ####
plot <- F
if(plot){
ggplot(data = dat, aes(x = po2, y = -depth, col = log(cpue_kg_km2+1))) +
  geom_point() +
  scale_colour_viridis_c(limits = c(0, 6), oob = scales::squish) +
  xlim(c(0, 16))
}

### Fit Breakpoint model to po2 ####
#Randomly assign folds (with specified seed)
seed <- 42 #872 #1929
set.seed(seed)
k_folds <- 10
dat$fold_ids <- sample(seq_len(k_folds), nrow(dat), replace = TRUE)
#Leave out only 10% of data
test_10 <- T
if(test_10){
  k_folds <- 2
 dat$fold_idsclust <- sample(1:2, size = nrow(dat), replace = T, prob = c(0.1, 0.9))
}
future::plan(future::multisession)
start <- matrix(c(20, 1.1), ncol = 1L) #sablefish
#start <- init_vals[["Sebastolobus altivelis"]][["m1"]][["start"]]

m1 <- sdmTMB_cv(cpue_kg_km2 ~ -1+year+breakpt(po2_s)+log_depth_scaled+log_depth_scaled2, 
             data = dat,
             time = NULL,
             reml = F,
             anisotropy = TRUE,
             spatiotemporal = FALSE,
             mesh=mesh,
             k_folds=k_folds,
             fold_ids = dat$fold_ids,
             family =tweedie(link="log"),
            future_globals = c("start"),
             control = sdmTMBcontrol(
               start = list(b_threshold = start)))

m1$models
m1$elpd
m1$fold_elpd


#### Plot fitted relationship ####

##### extract estimates ####
m1_pars <- m1$sd_report$par.fixed
m1_parnames <- names(m1_pars)
b_threshold <- m1_pars[grep("b_threshold", m1_parnames)]
##### plot ####
dat$thresh_pred <- exp(brkptfun(x = dat$po2_s, b_slope = b_threshold[1], b_thresh = b_threshold[2]))
ggplot(dat, aes(x = po2, y = thresh_pred, col = log(cpue_kg_km2 + 1))) +
  geom_point()
  #geom_jitter(width = 0.1, height = 0.01) +
  scale_colour_viridis_c(limits = c(0, 5), oob = scales::squish)+
  xlim(c(0,5))+
  ylab("Marginal Effect")
  

### Fit Eo estimation - po2 prime model ####
#Set starting parameters: 
start <- init_vals$"sablefish"$m2$start
#start <- init_vals$"Sebastolobus altivelis"$m2$start
#lower <- init_vals$"Sebastolobus altivelis"$m2$lower
#upper <- init_vals$"Sebastolobus altivelis"$m2$upper
m2 <- sdmTMB_cv(cpue_kg_km2 ~ -1+year+logistic(mi)+log_depth_scaled+log_depth_scaled2, 
             data = dat, 
             time = NULL,
             reml = F,
             anisotropy = TRUE,
             spatiotemporal = FALSE,
             mesh=mesh,
             family =tweedie(link="log"),
             k_folds=k_folds,
             fold_ids = dat$fold_ids,
           future_globals = c("start"),
             control = sdmTMBcontrol(
              start = list(b_threshold = start),
            #lower = list(b_threshold = lower),
             # upper = list(b_threshold = upper)
               newton_loops = 2,
           nlminb_loops=2))

m2$models
m2$elpd
m2$fold_elpd

#### Plot fitted relationshop ####
##### extract estimates ####

Eo <- getEo(model = m2)

##### plot ####
dat$po2_prime <- dat$po2 * exp(Eo * dat$invtemp)
dat$Eo_unconstrained <- exp(logfun(dat$po2_prime, model = m2, mi = T))
ggplot(dat, aes(x = po2_prime, y = Eo_unconstrained, col = log(cpue_kg_km2 + 1))) +
  geom_jitter(width = 0.1, height = 0.01) +
  scale_colour_viridis_c(limits = c(0, 6), oob = scales::squish) + 
  xlim(c(0,5))+
  ylab("Marginal Effect")
  
### Fit Eo estimation - po2 prime model (with prior) ####
## Set starting parameters:

#start <- init_vals$"Sebastolobus altivelis"$m2a$start
start <- matrix(c(1.4, 0.1, 50, 0.1), ncol=1)

prior <- normal(c(NA, NA, NA, 0.331), c(NA, NA, NA, 0.176))

m2a <- sdmTMB_cv(cpue_kg_km2 ~ -1+year+logistic(mi)+log_depth_scaled+log_depth_scaled2,
             data = dat, 
             time = NULL,
             reml = F,
             anisotropy = TRUE,
             spatiotemporal = FALSE,
             mesh=mesh,
             family =tweedie(link="log"),
             priors=sdmTMBpriors(threshold = prior),
             k_folds=k_folds,
             fold_ids = dat$fold_ids,
             future_globals = c("prior", "start"),
             control = sdmTMBcontrol(
              start = list(b_threshold = start),
            #   upper = list(b_threshold = upper),
             #  lower = list(b_threshold = lower),
               newton_loops = 3,
            nlminb_loops=2))

m2a$models
m2a$elpd
m2a$fold_elpd

#### Plot fitted relationship ####
##### extract estimates ####
Eo <- getEo(model = m2a)

##### plot ####
## Calculate po2 prime from parameter ##
dat$po2_prime <- dat$po2 * exp(Eo * dat$invtemp)

##### plot ####

dat$Eo_constrained <- exp(logfun(dat$po2_prime, model = m2a, mi = T))
ggplot(dat, aes(x = po2_prime, y = Eo_constrained, col = log(cpue_kg_km2 + 1))) +
  geom_jitter(width = 0.1, height = 0.01) +
  scale_colour_viridis_c(limits = c(0, 6), oob = scales::squish) +
  xlim(c(0,5))+
  ylab("Marginal Effect")
  


### Fit logistic po2 model ####
#Starting values (work for both all depths and depth-constraine)
start <- init_vals$sablefish$m3$start
start <- init_vals$"Sebastolobus altivelis"$m3$start
start <- matrix(c(-0.9, -0.6,9))

m3 <- sdmTMB_cv(cpue_kg_km2 ~ -1+year+logistic(po2_s)+log_depth_scaled+log_depth_scaled2,
             data = dat, 
             spatial = "on",
             mesh=mesh,
             anisotropy=T,
             reml=F,
             time=NULL,
             family =tweedie(link="log"),
             future_globals = c("start"),
             k_folds=k_folds,
             fold_ids = dat$fold_ids,
             control = sdmTMBcontrol(
               start = list(b_threshold=start),
             #  lower = list(b_threshold = lower),
             #  upper = list(b_threshold = upper)
               )
        )


m3$models
m3$sum_loglik
m3$fold_loglik


#### Plot fitted relationship ####
##### extract estimates ####
dat$logistic_pos <- exp(logfun(dat$po2_s, model = m3, mi = F))
##### plot ####
ggplot(dat, aes(x = po2, y = logistic_pos, col = log(cpue_kg_km2 + 1))) +
  geom_jitter(width = 0.1, height = 0.01) +
  scale_colour_viridis_c(limits = c(0, 6), oob = scales::squish)+
  xlim(c(0,5))+
  ylab("Marginal Effect")

### Fit null model ####
m4 <- sdmTMB_cv(cpue_kg_km2 ~ -1+year+log_depth_scaled+log_depth_scaled2, 
             data = dat, 
             spatial = "on",
             mesh=mesh,
             anisotropy=T,
             reml=F,
             time=NULL,
             k_folds=k_folds,
             fold_ids = dat$fold_ids,
             future_globals = c("start"),
             family =tweedie(link="log"),
             control = sdmTMBcontrol(
               newton_loops = 2,
               )
             )
m34models
m4sum_loglik
m4$fold_loglik

### Fit other alternative models ###
## Temperature ##

m5 <- sdmTMB_cv(cpue_kg_km2 ~ -1+year+log_depth_scaled+log_depth_scaled2+temp_s,
             data = dat, 
             spatial = "on",
             mesh=mesh,
             anisotropy=T,
             reml=F,
             time=NULL,
             k_folds=k_folds,
             fold_ids = dat$fold_ids,
             future_globals = c("start"),
             family =tweedie(link="log"),
             control = sdmTMBcontrol(
               newton_loops = 1,
             )
)

m5$models
m5$sum_loglik
m5$fold_loglik

##Oxygen only##
m6 <- sdmTMB_cv(cpue_kg_km2 ~ -1+year+log_depth_scaled+log_depth_scaled2+po2_s,
             data = dat, 
             spatial = "on",
             mesh=mesh,
             anisotropy=T,
             reml=F,
             k_folds=k_folds,
             fold_ids = dat$fold_ids,
             time=NULL,
             family =tweedie(link="log"),
             control = sdmTMBcontrol(newton_loops = 1,
             )
)

m6$models
m6$sum_loglik
m6$fold_loglik

## Temp and o2 ##
m7 <- sdmTMB_cv(cpue_kg_km2 ~ -1+year+log_depth_scaled+log_depth_scaled2+temp_s + po2_s,
             data = dat, 
             spatial = "on",
             mesh=mesh,
             anisotropy=T,
             reml=F,
             k_folds=k_folds,
             fold_ids = dat$fold_ids,
             time=NULL,
             family =tweedie(link="log"),
             control = sdmTMBcontrol(newton_loops = 1,
             )
)

m7$models
m7$sum_loglik
m7$fold_loglik

##Temp and o2 interaction ##
m8 <- sdmTMB_cv(cpue_kg_km2 ~ -1+year+log_depth_scaled+log_depth_scaled2+temp_s * po2_s,
             data = dat, 
             spatial = "on",
             mesh=mesh,
             anisotropy=T,
             reml=F,
             k_folds=k_folds,
             fold_ids = dat$fold_ids,
             time=NULL,
             family =tweedie(link="log"),
             control = sdmTMBcontrol(newton_loops = 1,
             )
)

m8$models
m8$sum_loglik
m8$fold_loglik

### Create log(prediction) (ie loss) table ###
## Make list of model names##
models <- c("breakpt-pO2", "Eo estimation and logistic po2 (no prior)", "Eo estimation and logistic po2' (prior)","logistic-pO2", "Null", "temp", "po2", "temp+po2", "temp * po2")
## Create table and add loss for each (totals) ##
compare<- as.data.frame(matrix(NA, ncol = 1, nrow =length(models), dimnames = list(models)))
compare[1,] <- m1$sum_loglik
compare[2,] <- m2$sum_loglik
compare[3,] <- m2a$sum_loglik
compare[4,] <- m3$sum_loglik
compare[5,] <- m4$sum_loglik
compare[6,] <- m5$sum_loglik
compare[7,] <- m6$sum_loglik
compare[8,] <- m7$sum_loglik
compare[9,] <- m8$sum_loglik

compare[1,] <- m1$fold_loglik[1]
compare[2,] <- m2$fold_loglik[1]
compare[3,] <- m2a$fold_loglik[1]
compare[4,] <- m3$fold_loglik[1]
compare[5,] <- m4$fold_loglik[1]
compare[6,] <- m5$fold_loglik[1]
compare[7,] <- m6$fold_loglik[1]
compare[8,] <- m7$fold_loglik[1]
compare[9,] <- m8$fold_loglik[1]

compare[1,2] <- m1$fold_loglik[1]
compare[2,2] <- m2$fold_loglik[1]
compare[3,2] <- m2a$fold_loglik[1]
compare[4,2] <- m3$fold_loglik[1]
compare[5,2] <- m4$fold_loglik[1]
compare[6,2] <- m5$fold_loglik[1]
compare[7,2] <- m6$fold_loglik[1]
compare[8,2] <- m7$fold_loglik[1]
compare[9,2] <- m8$fold_loglik[1]

compare$dLL <- abs(max(compare$V1)-(compare$V1))

## For each fold ##

save <- F
if(save){
save(m1,m2,m2a,m3,m4,m5,m6,m7,m8, file="models_thornyhead.R")
save(compare, compare2, file="thornyhead_model_selection.R")
}

## Calculate comparison ##


#### Plot marginal effects w/ +- SE using Tim's sdm_o2 code
## Breakpoint ##
# Make model output tidy #
tidy(m1,"ran_pars",conf.int = TRUE)
tidy(m1,"fixed", conf.int = TRUE)

# Create new data with everything set to 0 except for pO2 #
nd_po2 <- data.frame(po2_s = seq(min(dat$po2_s), max(dat$po2_s), length.out = 300), 
                     temp = 0,
                     log_depth_scaled = 0,
                     log_depth_scaled2 = 0,
                     year = as.factor(2010L)
                     )
nd_po2 <- convert_class(nd_po2)

# predict to new data #
p_o2 <- predict(m1, newdata = nd_po2, se_fit = TRUE, re_form = NA)

# plot predictions with uncertainty
z <- 1.645 # for 90% CI
p_o2$po2 <-back.convert(p_o2$po2_s, attr(dat$po2_s, "scaled:center"), attr(dat$po2_s, "scaled:scale"))
plot_breakpoint <- ggplot(p_o2, aes(x=po2, y=exp(est), 
                            ymin = exp(est - z *est_se), ymax = exp(est + z * est_se))) +
  geom_line() + geom_ribbon(alpha = 0.4) + 
  scale_x_continuous(limits = c(0, 3), expand = expansion(mult = c(0, 0.0))) +
  labs(x = "Partial Pressure of Oxygen (kPa)", y = bquote('Population Density'~(kg~km^-2)))

plot_breakpoint

## Constrained Eo ##
tidy(m2a,"ran_pars",conf.int = TRUE)
tidy(m2a,"fixed", conf.int = TRUE)

# predict on full dataframe # 
nd_po2 <- data.frame(po2 = seq(min(dat$po2), max(dat$po2), length.out = 300), 
                     invtemp = mean(dat$invtemp),
                     log_depth_scaled = 0,
                     log_depth_scaled2 = 0,
                     year = as.factor(2010L)
)
nd_po2 <- convert_class(nd_po2)
p_o2 <- predict(m2a, newdata = nd_po2, se_fit = TRUE, re_form = NA)

#Calculate po2 prime #
p_o2$prime <- p_o2$po2 * exp(Eo * p_o2$invtemp)

# Plot--ends up super squiggly #
ggplot(p_o2, aes(x=prime, y=exp(est),
                 ymin = exp(est - z *est_se), ymax = exp(est + z * est_se))) +
  geom_line() + geom_ribbon(alpha = 0.4) + 
  scale_y_continuous(limits = c(0, 75), expand = expansion(mult = c(0, 0.0))) +
  scale_x_continuous(limits = c(0, 1.5), expand = expansion(mult = c(0, 0.0))) +
  labs(x = "Partial Pressure of Oxygen (kPa)", y = bquote('Population Density'~(kg~km^-2)))

#Try to manually thin it to create a smoother line--this still makes a squiggly plot #
p_o2 <- arrange(p_o2, p_o2$prime)
ggplot(p_o2[seq(1, nrow(p_o2), 10), ], aes(x=prime, y=exp(est),
                                          ymin = exp(est - z *est_se), ymax = exp(est + z * est_se))) +
  geom_line() + geom_ribbon(alpha = 0.4) + 
  scale_y_continuous(limits = c(0, 75), expand = expansion(mult = c(0, 0.0))) +
  scale_x_continuous(limits = c(0, 1.5), expand = expansion(mult = c(0, 0.0))) +
  labs(x = "Partial Pressure of Oxygen (kPa)", y = bquote('Population Density'~(kg~km^-2)))

##Option 2: creating new dataframe with smoothed po2 and invtemp #
# Back-calculate pO2 prime, po2, and invtemp values #
Eo <- getEo(model = m2a)
dat$po2_prime <- dat$po2 * exp(Eo * dat$invtemp)
prime_preds <- data.frame(prime =seq(min(dat$po2_prime), max(dat$po2_prime), length.out=300))
prime_preds$po2 <- seq(min(dat$po2), max(dat$po2), length.out=300)
prime_preds$invtemp <- log(prime_preds$prime/prime_preds$po2)/Eo
prime_preds$test <- prime_preds$po2 * exp(Eo * prime_preds$invtemp)

nd_po2 <- data.frame(po2 = prime_preds$po2, 
                     invtemp = prime_preds$invtemp,
                     log_depth_scaled = 0,
                     log_depth_scaled2 = 0,
                     year = as.factor(2010L)
)
nd_po2 <- convert_class(nd_po2)

# predict to new data--this Rbombs#
p_o2 <- predict(m2a, newdata = nd_po2, se_fit = TRUE, re_form = NA)

ggplot(p_o2, aes(x=prime, y=exp(est),
                                           ymin = exp(est - z *est_se), ymax = exp(est + z * est_se))) +
  geom_line() + geom_ribbon(alpha = 0.4) + 
  scale_y_continuous(limits = c(0, 75), expand = expansion(mult = c(0, 0.0))) +
  scale_x_continuous(limits = c(0, 1.5), expand = expansion(mult = c(0, 0.0))) +
  labs(x = "Partial Pressure of Oxygen (kPa)", y = bquote('Population Density'~(kg~km^-2)))

## Plot logistic ##
# Make model output tidy #
tidy(m3,"ran_pars",conf.int = TRUE)
tidy(m3,"fixed", conf.int = TRUE)

# Create new data with everything set to 0 except for pO2 #
nd_po2 <- data.frame(po2_s = seq(min(dat$po2_s), max(dat$po2_s), length.out = 300), 
                     temp = 0,
                     log_depth_scaled = 0,
                     log_depth_scaled2 = 0,
                     year = as.factor(2010L)
)
nd_po2 <- convert_class(nd_po2)

# predict to new data #
p_o2 <- predict(m3, newdata = nd_po2, se_fit = TRUE, re_form = NA)

# plot predictions with uncertainty
z <- 1.645 # for 90% CI
p_o2$po2 <-back.convert(p_o2$po2_s, attr(dat$po2_s, "scaled:center"), attr(dat$po2_s, "scaled:scale"))
plot_logistic <- ggplot(p_o2, aes(x=po2, y=exp(est), 
                                    ymin = exp(est - z *est_se), ymax = exp(est + z * est_se))) +
  geom_line() + geom_ribbon(alpha = 0.4) + 
  scale_y_continuous(limits = c(0, 250), expand = expansion(mult = c(0, 0.0))) +
  scale_x_continuous(limits = c(0, 1), expand = expansion(mult = c(0, 0.0))) +
  labs(x = "Partial Pressure of Oxygen (kPa)", y = bquote('Population Density'~(kg~km^-2)))

plot_logistic

multipanel <- gridExtra::grid.arrange(breakpoint, logistic, Eo_constrained, nrow = 3, ncol = 1)
ggsave("plots/marginal_effects.png", multipanel, width = 6, height = 3, units = "in", device = "png")







##### Older versions, ignore ####
### Plot depth effects  ####
## Set ggplot theme ##
theme_set(theme_bw(base_size = 30))
theme_update(panel.grid.major = element_blank(), panel.grid.minor = element_blank())
## Internal Eo mode ##
# Extract parameters another way for internal Eo model #
par_estimates <- as.data.frame(tidy(m2, conf.int = TRUE, effects="fixed"))
par_estimates_rand <- as.data.frame(tidy(m2, conf.int = TRUE, effects="ran_pars"))
par_estimates <- bind_rows(par_estimates, par_estimates_rand)
pars <- par_estimates[1:2]
pars <- pivot_wider(pars, names_from=term, values_from=estimate)
# Calculate temp-corrected po2 from Eo value #
dat$mi_pred <- dat$po2*exp(pars$"mi-Eo"* dat$invtemp)
# Calculate effect of MI # Tim note: probably better to use the logfun function above to calculate the log effect size from po2prime or po2.  
dat$mi_effect <- pars$"mi-smax" * (1 / (1 + exp(-log(19) * (dat$mi_pred - pars$`mi-s50`) / pars$"mi-delta")) - 1) # this is exponentiated. 
# Calculate combined depth effect #
dat$depth_effect_combined <- (dat$log_depth_scaled*pars$log_depth_scaled)+(dat$log_depth_scaled2*pars$log_depth_scaled2)

## For breakpt(o2) model ##
par_estimates2 <- as.data.frame(tidy(m1, conf.int = TRUE, effects="fixed"))
par_estimates_rand2 <- as.data.frame(tidy(m1, conf.int = TRUE, effects="ran_pars"))
par_estimates2 <- bind_rows(par_estimates2, par_estimates_rand2)
pars2 <- par_estimates2[1:2]
pars2 <- pivot_wider(pars2, names_from=term, values_from=estimate)
# alculate depth effects #
dat$depth_effect_combined2 <- (dat$log_depth_scaled*pars2$log_depth_scaled)+(dat$log_depth_scaled2*pars2$log_depth_scaled2)
# Calculate po2 effect #
dat$po2_effect <- ifelse(dat$po2_s>pars2$`po2_s-breakpt`,pars2$`po2_s-breakpt`*pars2$`po2_s-slope`,dat$`po2_s`*pars2$`po2_s-slope`)

# Plot depth effects compared between model #
ggplot(dat, aes(y=depth_effect_combined, x=depth))+geom_line(size=1.3)+geom_line(dat, mapping=aes(y=depth_effect_combined2, x=depth), color="red", size=1.3)+ylab("Depth Effect Combined")
ggplot(subset(dat, depth>450), aes(y=depth_effect_combined, x=depth))+geom_point()+geom_point(subset(dat, depth>450), mapping=aes(y=depth_effect_combined2, x=depth), color="red")+ylab("Depth Effect Combined")

# Plot metabolic index vs depth #
ggplot(dat, aes(y=depth, x=mi_pred))+geom_point(aes(color=log(cpue_kg_km2)))+xlab("pO2'")+theme(legend.position=c(0.8, 0.8))

#Plot oxygen vs depth
ggplot(dat, aes(y=-depth, x=po2))+geom_point(aes(color=log(cpue_kg_km2)))+xlab("pO2")+theme(legend.position=c(0.7, 0.3))+scale_colour_viridis_c()

#Save models
save(m1,m2,m2a,m3,m4,m5,m6,m7,m8, file="models_petralesole.RDS")

#Calculate percent below
below <- subset(dat, dat$po2_prime < 1.45 & dat$cpue_kg_km2>0)
below2 <- subset(dat,  dat$cpue_kg_km2>0)
below <- subset(dat, po2_s < -1.09 & dat$cpue_kg_km2>0)
